complete level editor and removed warnings and overall structure

+++ code/Pacman/BlueEffect.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "buffeffect.h"
+
 class BlueEffect :
 	public BuffEffect
 {
@@ -7,6 +8,5 @@ public:
 	BlueEffect();
 	~BlueEffect();
 
-	void update(float dt);
+	void update(double dt);
 };
-
